requirements.txt
# Library
pyrogram
python-telegram-bot
# Other
aiohttp
alphabet_detector
bleach
bs4
deezloader
demjson
emoji
feedparser
future
git+https://github.com/AkaruiAikara/python-markdown2.git
google_trans_new
gtts
jikanpy
lxml
nekobin
nekos.py
pillow
psutil
psycopg2-binary
pythonping
pytz
requests
search_engine_parser
spamwatch
speedtest-cli>=2.0.2
sqlalchemy
telegraph
tgcrypto
tswift
wget
wikipedia
